Added missing component prop
