+++ app/client/src/components/pages/Community.Tabs.Overview.js
@@ -35,7 +35,10 @@ import VirtualizedList from 'components/shared/VirtualizedList';
 // contexts
 import { useLayers } from 'contexts/Layers';
 import { LocationSearchContext } from 'contexts/locationSearch';
-import { useServicesContext } from 'contexts/LookupFiles';
+import {
+  useCharacteristicsByGroupContext,
+  useServicesContext,
+} from 'contexts/LookupFiles';
 // utilities
 import {
   useCyanWaterbodies,
@@ -436,6 +439,8 @@ function MonitoringAndSensorsTab({
     useMonitoringLocations();
   const { cyanWaterbodies, cyanWaterbodiesStatus } = useCyanWaterbodies();
 
+  const characteristicsByGroup = useCharacteristicsByGroupContext();
+
   const allMonitoringAndSensors = [
     ...streamgages,
     ...monitoringLocations,
@@ -656,6 +661,7 @@ function MonitoringAndSensorsTab({
           >
             <div css={accordionContentStyles}>
               <WaterbodyInfo
+                characteristicsByGroup={characteristicsByGroup}
                 type={monitoringType}
                 feature={feature}
                 services={services}
@@ -667,6 +673,7 @@ function MonitoringAndSensorsTab({
       }
     },
     [
+      characteristicsByGroup,
       accordionItemToggleHandlers,
       expandedRows,
       filteredMonitoringAndSensors,
